Update README.md
BLOCKCHAIN AS A TECHNOLOGY HAS ENTERED ITS CAMBRIAN STAGE.A glut of investor interest has led to an explosion in the technical diversity of projects now underway. Money which supports the industry not directed to the right direction case fragmentation in the industry. At the heart of this burgeoning inovation is the AFKOIN-BLOCKCHAIN-SPACE. Though the impulse to compete is at a peak so too is the impulse to colaboraborate